* Meta field add functions * put_constant() and put_string() * create empty binary file when a RAW field is added.

+++ test/add_raw.c
@@ -13,9 +13,10 @@ int main(void)
 {
   const char* filedir = __TEST__ "dirfile";
   const char* format = __TEST__ "dirfile/format";
+  const char* data = __TEST__ "dirfile/data";
 
   DIRFILE* D = dirfile_open(filedir, GD_RDWR | GD_CREAT);
-  dirfile_add_raw(D, "new", GD_UINT8, 2, 0);
+  dirfile_add_raw(D, "data", GD_UINT8, 2, 0);
   int error = get_error(D);
 
   /* check */
@@ -23,6 +24,8 @@ int main(void)
 
   dirfile_close(D);
 
+  if (unlink(data))
+    return 1;
   unlink(format);
   rmdir(filedir);
